+++ CHANGELOG.md
@@ -14,6 +14,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ## [2.18.0] - 2019-07-23
 
 ### Added
+
 - MenuItems can have icon now by receiving the `iconProps` and `iconToTheRight` props
 
 ## [2.17.2] - 2019-07-16